move server to web/ lib
